git-svn-id: http://crux-framework.googlecode.com/svn@3433 a5d2bbaa-053c-11de-b17c-0f1ef23b492c

+++ branches/4.1.0/CruxHangout/src/org/cruxframework/crux/hangout/client/data/StateMetadata.java
@@ -33,11 +33,11 @@ public class StateMetadata extends JavaScriptObject
 		return this.value;
 	}-*/;
 
-	public final native long getTimestamp()/*-{
+	public final native double getTimestamp()/*-{
 		return this.timestamp;
 	}-*/;
 
-	public final native long getTimeDiff()/*-{
+	public final native double getTimeDiff()/*-{
 		return this.timediff;
 	}-*/;
 }
